pyproject.toml
[tool.poetry]
name = "ambiente-inteligente"
version = "0.1.0"
description = ""
authors = ["None"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
protobuf = "^5.29.2"
streamlit = "^1.41.1"
python-dotenv = "^1.0.1"
paho-mqtt = "^2.1.0"
python-box = "^7.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
